+++ src/components/domain/random/RandomContent.tsx
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+const RandomContent = styled.div`
+  position: relative;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 36px;
+  padding: 0 36px 0;
+`;
+
+export default RandomContent;
