+++ src/pages/random/voice/[idx].tsx
@@ -4,14 +4,13 @@ import useStorage from '~/hooks/useStorage';
 import PostHeader from '~/components/domain/question/PostHeader';
 import { IQuestionDetail } from '~/types/question';
 import Recorder from '~/components/domain/question/Recorder';
-import styled from '@emotion/styled';
-import Button from '~/components/base/Button';
 import PageContainer from '~/components/common/PageContainer';
-import Icon from '~/components/base/Icon';
 import { isValidRandomType } from '~/utils/helper/validation';
 import { isString } from '~/utils/helper/checkType';
 import { RANDOM_LOCAL_KEY } from '~/utils/constant/random';
 import Article from '~/components/common/Article';
+import MoveButtons from '~/components/common/MoveButtons';
+import RandomContent from '~/components/domain/random/RandomContent';
 
 const DUMMY = 1;
 const DUMMY_QUESTION = {} as IQuestionDetail;
@@ -36,15 +35,6 @@ const RandomVoice = () => {
   const handleSpeechEnd = (recordRefId: number) => {
     if (recordRefId !== recordRefLastId.current) return;
     recordRef.current?.click();
-  };
-
-  const handlePrev = () => {
-    if (!curQuestion || !questions || curQuestion.idx === 1) {
-      alert('첫 번째 질문입니다.');
-      return;
-    }
-
-    move(curQuestion.idx - 1);
   };
 
   const handleNext = async () => {
@@ -117,20 +107,14 @@ const RandomVoice = () => {
               writer={curQuestion.nickname}
             />
 
-            <PostContent>
+            <RandomContent>
               <Recorder ref={recordRef} limit={1} key={curQuestion.idx} />
-
-              <Buttons>
-                <PrevButton buttonType="borderGray" onClick={handlePrev}>
-                  <Icon name="ArrowLeft" size="20" color="gray600" />
-                  <span>이전 질문</span>
-                </PrevButton>
-                <NextButton buttonType="borderGray" onClick={handleNext}>
-                  <span> 다음 질문</span>
-                  <Icon name="ArrowRight" size="20" color="gray600" />
-                </NextButton>
-              </Buttons>
-            </PostContent>
+              <MoveButtons
+                disabledPrev={!curQuestion || !questions || curQuestion.idx === 1}
+                onPrev={() => move(curQuestion.idx - 1)}
+                onNext={handleNext}
+              />
+            </RandomContent>
           </Article>
         </PageContainer>
       )}
@@ -187,32 +171,3 @@ function isValidStoredValue(value: unknown) {
 function isValidType(type: unknown) {
   return isValidRandomType(type) && type === 'voice';
 }
-
-const PostContent = styled.div`
-  position: relative;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 36px;
-  padding: 0 36px 0;
-`;
-
-const Buttons = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  margin-top: 56px;
-`;
-
-const PrevButton = styled(Button)`
-  display: flex;
-  align-items: center;
-  padding-left: 16px;
-`;
-
-const NextButton = styled(Button)`
-  display: flex;
-  align-items: center;
-  padding-right: 16px;
-`;
